requirements.txt
dpkt==1.9.4
python-pcapng==1.0